VirtualMachineEndpoints API
This patch defines the VirtualMachineEndpoints API to support
manually specifying the endpoint information for a
VirtualMachineService resource.

This is useful to support Antrea's NodePortLocal feature as
well as when DevOps users on Supervisor want to control which
IP a VirtualMachineService uses for a given port.

The Makefile was also fixed for the new version of controller-
gen.